Make also the pan-property float (saves scaling and yields better resolution)
Original commit message from CVS:
* gst/audiofxgood/audiopanorama.c: (gst_audio_panorama_class_init),
(gst_audio_panorama_set_property),
(gst_audio_panorama_get_property),
(gst_audio_panorama_transform_m2s_int),
(gst_audio_panorama_transform_s2s_int),
(gst_audio_panorama_transform_m2s_float),
(gst_audio_panorama_transform_s2s_float):
* gst/audiofxgood/audiopanorama.h:
* tests/check/elements/audiopanorama.c: (GST_START_TEST):
Make also the pan-property float (saves scaling and yields better
resolution)

+++ gst/audiofx/audiopanorama.h
@@ -41,7 +41,7 @@ typedef void (*GstAudioPanoramaProcessFunc)(GstAudioPanorama*, guint8*, guint8*,
 struct _GstAudioPanorama {
   GstBaseTransform element;
 
-  gint panorama;
+  gfloat panorama;
   
   /* < private > */
   GstAudioPanoramaProcessFunc process;
